requirements.txt
pandas
requests
python-dotenv
spacy
openai
